requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=1.7.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
attrs=25.3.0=pypi_0
autopep8=2.3.2=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.0=pypi_0
colorama=0.4.6=pypi_0
datasets=3.6.0=pypi_0
device-smi=0.4.1=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
faiss-cpu=1.11.0=pypi_0
filelock=3.18.0=pypi_0
flask=3.1.1=pypi_0
frozenlist=1.6.0=pypi_0
fsspec=2025.3.0=pypi_0
gptqmodel=2.2.0+cu121torch2.5=pypi_0
greenlet=3.2.2=pypi_0
h11=0.16.0=pypi_0
hf-transfer=0.1.9=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.31.2=pypi_0
idna=3.10=pypi_0
iniconfig=2.1.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
joblib=1.5.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.25=pypi_0
langchain-core=0.3.60=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langsmith=0.3.42=pypi_0
libffi=3.4.4=hd77b12b_1
logbar=0.0.4=pypi_0
markupsafe=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
networkx=3.4.2=pypi_0
numpy=2.2.5=pypi_0
openai=1.79.0=pypi_0
openssl=3.0.16=h3f729d1_0
optimum=1.25.3=pypi_0
orjson=3.10.18=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_2
pluggy=1.6.0=pypi_0
propcache=0.3.1=pypi_0
protobuf=6.31.0=pypi_0
psutil=7.0.0=pypi_0
pyarrow=20.0.0=pypi_0
pycodestyle=2.13.0=pypi_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pypdf2=3.0.1=pypi_0
pytest=8.3.5=pypi_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
random-word=1.0.13=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.3=pypi_0
sentence-transformers=4.1.0=pypi_0
setuptools=78.1.1=py311haa95532_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.41=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenicer=0.0.4=pypi_0
tokenizers=0.21.1=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchvision=0.20.1+cu121=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.51.3=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py311haa95532_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h4754444_1
yarl=1.20.0=pypi_0
zlib=1.2.13=h8cc25b3_1
zstandard=0.23.0=pypi_0
